// apps/pc/src/generated/models/Pl_contractsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_contractspl_contractstatuscode = {
  100000000: '締結済み',
  100000001: '終了'
} as const;
export type Pl_contractspl_contractstatuscode = keyof typeof Pl_contractspl_contractstatuscode;
export const Pl_contractsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_contractsstatecode = keyof typeof Pl_contractsstatecode;
export const Pl_contractsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_contractsstatuscode = keyof typeof Pl_contractsstatuscode;

export interface Pl_contractsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_autorenew?: boolean;
  pl_contractid: string;
  pl_contractstatuscode?: Pl_contractspl_contractstatuscode;
  "pl_ContractTypeLookup@odata.bind"?: string;
  pl_decisiondate?: string;
  pl_enddate?: string;
  pl_link?: string;
  pl_name?: string;
  pl_noticedate?: string;
  "pl_PartnerLookup@odata.bind"?: string;
  pl_registrationkey?: string;
  statecode: Pl_contractsstatecode;
  statuscode?: Pl_contractsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_contracts extends Pl_contractsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_autorenewname?: string;
  pl_contractstatuscodename?: string;
  pl_contracttypelookupname?: string;
  pl_partnerlookupname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_contracttypelookup?: object;
  _pl_contracttypelookup_value?: string;
  pl_partnerlookup?: object;
  _pl_partnerlookup_value?: string;
}